switch back to f32

+++ libs/fft/fftsg_f32.cpp
@@ -1,4 +1,4 @@
-using value_t = double;
+using value_t = float;
 
 /*
 Fast Fourier/Cosine/Sine Transform
